:recycle: Move files and improve file naming and structure
