+++ src/openpersonen/api/data_classes/converters/__init__.py
@@ -1,0 +1,1 @@
+from .ingeschreven_persoon import *  # noqa
