+++ src/openpersonen/api/data_classes/kind.py
@@ -1,15 +1,11 @@
 from dataclasses import dataclass
-from datetime import datetime
 
 from django.conf import settings
 
-import xmltodict
-from dateutil.relativedelta import relativedelta
-
+from openpersonen.api.demo_models import Persoon as PersoonDemoModel
 from openpersonen.api.models import StufBGClient
-from openpersonen.api.testing_models import Persoon as PersoonModel
-from openpersonen.api.utils import convert_empty_instances
 
+from .converters.kind import get_client_instance_dict, get_model_instance_dict
 from .in_onderzoek import KindInOnderzoek
 from .persoon import Persoon
 
@@ -19,215 +15,26 @@ class Kind(Persoon):
     leeftijd: int
     inOnderzoek: KindInOnderzoek
 
-    @staticmethod
-    def get_client_instance_dict(response):
-        dict_object = xmltodict.parse(response.content)
-
-        antwoord_dict_object = dict_object["soapenv:Envelope"]["soapenv:Body"][
-            "ns:npsLa01"
-        ]["ns:antwoord"]["ns:object"]["ns:inp.heeftAlsKinderen"]["ns:gerelateerde"]
-
-        kind_dict = {
-            "burgerservicenummer": antwoord_dict_object["ns:inp.bsn"],
-            "geheimhoudingPersoonsgegevens": True,
-            "naam": {
-                "geslachtsnaam": antwoord_dict_object["ns:geslachtsnaam"],
-                "voorletters": antwoord_dict_object["ns:voorletters"],
-                "voornamen": antwoord_dict_object["ns:voornamen"],
-                "voorvoegsel": antwoord_dict_object["ns:voorvoegselGeslachtsnaam"],
-                "inOnderzoek": {
-                    "geslachtsnaam": bool(antwoord_dict_object["ns:geslachtsnaam"]),
-                    "voornamen": bool(antwoord_dict_object["ns:voornamen"]),
-                    "voorvoegsel": bool(
-                        antwoord_dict_object["ns:voorvoegselGeslachtsnaam"]
-                    ),
-                    "datumIngangOnderzoek": {
-                        "dag": 0,
-                        "datum": "string",
-                        "jaar": 0,
-                        "maand": 0,
-                    },
-                },
-            },
-            "geboorte": {
-                "datum": {
-                    "dag": int(
-                        antwoord_dict_object["ns:geboortedatum"][
-                            settings.DAY_START : settings.DAY_END
-                        ]
-                    ),
-                    "datum": antwoord_dict_object["ns:geboortedatum"],
-                    "jaar": int(
-                        antwoord_dict_object["ns:geboortedatum"][
-                            settings.YEAR_START : settings.YEAR_END
-                        ]
-                    ),
-                    "maand": int(
-                        antwoord_dict_object["ns:geboortedatum"][
-                            settings.MONTH_START : settings.MONTH_END
-                        ]
-                    ),
-                },
-                "land": {
-                    "code": "string",
-                    "omschrijving": antwoord_dict_object["ns:inp.geboorteLand"],
-                },
-                "plaats": {
-                    "code": "string",
-                    "omschrijving": antwoord_dict_object["ns:inp.geboorteplaats"],
-                },
-                "inOnderzoek": {
-                    "datum": True,
-                    "land": True,
-                    "plaats": True,
-                    "datumIngangOnderzoek": {
-                        "dag": 0,
-                        "datum": "string",
-                        "jaar": 0,
-                        "maand": 0,
-                    },
-                },
-            },
-            "leeftijd": relativedelta(
-                datetime.now(),
-                datetime.strptime(antwoord_dict_object["ns:geboortedatum"], "%Y%m%d"),
-            ).years,
-            "inOnderzoek": {
-                "burgerservicenummer": bool(antwoord_dict_object["ns:inp.bsn"]),
-                "datumIngangOnderzoek": {
-                    "dag": 0,
-                    "datum": "string",
-                    "jaar": 0,
-                    "maand": 0,
-                },
-            },
-        }
-
-        convert_empty_instances(kind_dict)
-
-        return kind_dict
-
-    @staticmethod
-    def get_model_instance_dict(kind):
-
-        kind_dict = {
-            "burgerservicenummer": kind.burgerservicenummer_kind,
-            "geheimhoudingPersoonsgegevens": True,
-            "naam": {
-                "geslachtsnaam": kind.geslachtsnaam_kind,
-                "voorletters": "string",
-                "voornamen": kind.voornamen_kind,
-                "voorvoegsel": kind.voorvoegsel_geslachtsnaam_kind,
-                "inOnderzoek": {
-                    "geslachtsnaam": bool(kind.geslachtsnaam_kind),
-                    "voornamen": bool(kind.voornamen_kind),
-                    "voorvoegsel": bool(kind.voorvoegsel_geslachtsnaam_kind),
-                    "datumIngangOnderzoek": {
-                        "dag": 0,
-                        "datum": "string",
-                        "jaar": 0,
-                        "maand": 0,
-                    },
-                },
-            },
-            "geboorte": {
-                "datum": {
-                    "dag": int(
-                        kind.geboortedatum_kind[settings.DAY_START : settings.DAY_END]
-                    )
-                    if kind.geboortedatum_kind
-                    else 0,
-                    "datum": kind.geboortedatum_kind,
-                    "jaar": int(
-                        kind.geboortedatum_kind[settings.YEAR_START : settings.YEAR_END]
-                    )
-                    if kind.geboortedatum_kind
-                    else 0,
-                    "maand": int(
-                        kind.geboortedatum_kind[
-                            settings.MONTH_START : settings.MONTH_END
-                        ]
-                    )
-                    if kind.geboortedatum_kind
-                    else 0,
-                },
-                "land": {
-                    "code": "string",
-                    "omschrijving": kind.geboorteland_kind,
-                },
-                "plaats": {
-                    "code": "string",
-                    "omschrijving": kind.geboorteplaats_kind,
-                },
-                "inOnderzoek": {
-                    "datum": True,
-                    "land": True,
-                    "plaats": True,
-                    "datumIngangOnderzoek": {
-                        "dag": 0,
-                        "datum": "string",
-                        "jaar": 0,
-                        "maand": 0,
-                    },
-                },
-            },
-            "leeftijd": relativedelta(
-                datetime.now(),
-                datetime.strptime(kind.geboortedatum_kind, "%Y%m%d"),
-            ).years
-            if kind.geboortedatum_kind
-            else 0,
-            "inOnderzoek": {
-                "burgerservicenummer": bool(kind.burgerservicenummer_kind),
-                "datumIngangOnderzoek": {
-                    "dag": int(
-                        kind.datum_ingang_onderzoek[
-                            settings.DAY_START : settings.DAY_END
-                        ]
-                    )
-                    if kind.datum_ingang_onderzoek
-                    else 0,
-                    "datum": kind.datum_ingang_onderzoek,
-                    "jaar": int(
-                        kind.datum_ingang_onderzoek[
-                            settings.YEAR_START : settings.YEAR_END
-                        ]
-                    )
-                    if kind.datum_ingang_onderzoek
-                    else 0,
-                    "maand": int(
-                        kind.datum_ingang_onderzoek[
-                            settings.MONTH_START : settings.MONTH_END
-                        ]
-                    )
-                    if kind.datum_ingang_onderzoek
-                    else 0,
-                },
-            },
-        }
-
-        return kind_dict
-
     @classmethod
     def list(cls, bsn):
         class_instances = []
         if getattr(settings, "USE_STUF_BG_DATABASE", False):
-            instances = PersoonModel.objects.get(
+            instances = PersoonDemoModel.objects.get(
                 burgerservicenummer_persoon=bsn
             ).kind_set.all()
             for instance in instances:
-                instance_dict = cls.get_model_instance_dict(instance)
+                instance_dict = get_model_instance_dict(instance)
                 class_instances.append(cls(**instance_dict))
         return class_instances
 
     @classmethod
     def retrieve(cls, bsn, id):
         if getattr(settings, "USE_STUF_BG_DATABASE", False):
-            instance = PersoonModel.objects.get(
+            instance = PersoonDemoModel.objects.get(
                 burgerservicenummer_persoon=bsn
             ).kind_set.get(burgerservicenummer_kind=id)
-            instance_dict = cls.get_model_instance_dict(instance)
+            instance_dict = get_model_instance_dict(instance)
         else:
             response = StufBGClient.get_solo().get_kind(bsn)
-            instance_dict = cls.get_client_instance_dict(response)
+            instance_dict = get_client_instance_dict(response)
         return cls(**instance_dict)
